feat(medication): 药品库存支持小数数量(DECIMAL)并优化家属端UI
- 药箱库存字段 total_quantity/remaining_quantity 改为 DECIMAL(12,3) 支持小数
- 新增 Flyway 迁移 V5__use_decimal_inventory_quantities.sql
- 同步更新实体、DTO、Mapper 及 Plan/MedicineBox 服务逻辑
- 家属端仪表盘/通知样式优化，新增 timeUtils 时间工具

+++ backend/src/main/java/com/example/backend/model/dto/AddMedicineRequest.java
@@ -5,6 +5,7 @@ import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 /**
  * 添加药品到药箱请求DTO
@@ -49,7 +50,7 @@ public class AddMedicineRequest {
      * 总数量（如30片、60片）
      */
     @NotNull(message = "总数量不能为空")
-    private Integer totalQuantity;
+    private BigDecimal totalQuantity;
 
     /**
      * 备注说明（最长500字符）
